backend/requirements.txt
Flask==2.2.5            # Core Flask framework
Flask-Cors==3.0.10      # Handle Cross-Origin Resource Sharing (optional, if needed)
pandas==2.1.3           # For exporting results to Excel
openpyxl==3.1.2         # To support Excel (.xlsx) file handling with pandas


aiodns==3.2.0
aiofiles==24.1.0
aiohttp==3.11.7
aiomultiprocess==0.9.1
aiosqlite==0.20.0
beautifulsoup4==4.12.3
censys==2.2.16
certifi==2024.8.30
dnspython==2.7.0
fastapi==0.115.5
lxml==5.3.0
netaddr==1.3.0
ujson==5.10.0
playwright==1.49.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
retrying==1.3.4
shodan==1.31.0
slowapi==0.1.9
uvicorn==0.32.1
uvloop==0.21.0; platform_system != "Windows"
winloop==0.1.6; platform_system == "Windows"
